1st version of LED toggling
I have modified many macros to function so to adapt the function pointer argument in GPIO_ClkMap and GPIO_Reset structures.
I added a main_gpio.c where I tested the driver API for a toggling demo, and delay with CYCYCLE = 5e5
The delay corresponds to 433 ms (see the captured image)
Also we correct the sigrok software for the udev rules.